VZ-6956.  Remove obsolete tests and refactor some common code (#4027)
* skip-at Remove obsolete rollingupdate/scheduling tests, refactor utils to common pkg
