+++ tests/e2e/ha/monitor/monitor_suite_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 	"github.com/verrazzano/verrazzano/pkg/test/framework"
-	"github.com/verrazzano/verrazzano/tests/e2e/ha"
+	"github.com/verrazzano/verrazzano/tests/e2e/pkg/ha"
 	"testing"
 )
