--- tests/e2e/ha/scheduling/scheduling_suite_test.go
@@ -1,15 +1,0 @@
-// Copyright (c) 2022, Oracle and/or its affiliates.
-// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-
-package scheduling
-
-import (
-	"github.com/onsi/ginkgo/v2"
-	"github.com/onsi/gomega"
-	"testing"
-)
-
-func TestScheduling(t *testing.T) {
-	gomega.RegisterFailHandler(ginkgo.Fail)
-	ginkgo.RunSpecs(t, "Worker node scheduling")
-}
