+++ tests/e2e/pkg/ha/signal.go
